DB int欄位改成bigint，C#連動修正

+++ BioMedDocManager/Controllers/AppActionController.cs
@@ -129,7 +129,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Edit =======================
-        public async Task<IActionResult> Edit([FromRoute] int? id)
+        public async Task<IActionResult> Edit([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -151,7 +151,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit([FromRoute] int? id, AppAction posted)
+        public async Task<IActionResult> Edit([FromRoute]  long? id, AppAction posted)
         {
             if (posted == null || id != posted.AppActionId)
             {
@@ -193,7 +193,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Details =======================
-        public async Task<IActionResult> Details([FromRoute] int? id)
+        public async Task<IActionResult> Details([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -267,7 +267,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Delete =======================
-        public async Task<IActionResult> Delete([FromRoute] int? id)
+        public async Task<IActionResult> Delete([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -340,7 +340,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Delete([FromRoute] int? id, AppAction posted)
+        public async Task<IActionResult> Delete([FromRoute]  long? id, AppAction posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.AppActionId)
             {
