+++ BioMedDocManager/Controllers/MenuItemController.cs
@@ -138,7 +138,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Edit =======================
-        public async Task<IActionResult> Edit([FromRoute] int? id)
+        public async Task<IActionResult> Edit([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -165,7 +165,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit([FromRoute] int? id, MenuItem posted)
+        public async Task<IActionResult> Edit([FromRoute]  long? id, MenuItem posted)
         {
             if (posted == null || id != posted.MenuItemId)
             {
@@ -218,7 +218,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Details =======================
-        public async Task<IActionResult> Details([FromRoute] int? id)
+        public async Task<IActionResult> Details([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -249,7 +249,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Delete =======================
-        public async Task<IActionResult> Delete([FromRoute] int? id)
+        public async Task<IActionResult> Delete([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -291,7 +291,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Delete([FromRoute] int? id, MenuItem posted)
+        public async Task<IActionResult> Delete([FromRoute]  long? id, MenuItem posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.MenuItemId)
             {
