+++ BioMedDocManager/Controllers/ResourceController.cs
@@ -133,7 +133,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Edit =======================
-        public async Task<IActionResult> Edit([FromRoute] int? id)
+        public async Task<IActionResult> Edit([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -154,7 +154,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit([FromRoute] int? id, Resource posted)
+        public async Task<IActionResult> Edit([FromRoute]  long? id, Resource posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.ResourceId)
             {
@@ -203,7 +203,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Details =======================
-        public async Task<IActionResult> Details([FromRoute] int? id)
+        public async Task<IActionResult> Details([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -303,7 +303,7 @@ namespace BioMedDocManager.Controllers
 
 
         // ======================= Delete =======================
-        public async Task<IActionResult> Delete([FromRoute] int? id)
+        public async Task<IActionResult> Delete([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -405,7 +405,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Delete([FromRoute] int? id, Resource posted)
+        public async Task<IActionResult> Delete([FromRoute]  long? id, Resource posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.ResourceId)
             {
