+++ BioMedDocManager/Controllers/RoleController.cs
@@ -130,7 +130,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Edit =======================
-        public async Task<IActionResult> Edit([FromRoute] int? id)
+        public async Task<IActionResult> Edit([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -153,7 +153,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit([FromRoute] int? id, Role posted)
+        public async Task<IActionResult> Edit([FromRoute]  long? id, Role posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.RoleId)
             {
@@ -197,7 +197,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Details =======================
-        public async Task<IActionResult> Details([FromRoute] int? id)
+        public async Task<IActionResult> Details([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -242,7 +242,7 @@ namespace BioMedDocManager.Controllers
         }
 
         // ======================= Delete =======================
-        public async Task<IActionResult> Delete([FromRoute] int? id)
+        public async Task<IActionResult> Delete([FromRoute]  long? id)
         {
             if (id.GetValueOrDefault() <= 0)
             {
@@ -309,7 +309,7 @@ namespace BioMedDocManager.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed([FromRoute] int? id, Role posted)
+        public async Task<IActionResult> DeleteConfirmed([FromRoute]  long? id, Role posted)
         {
             if (posted == null || id.GetValueOrDefault() <= 0 || id != posted.RoleId)
             {
